+++ HeapNode.java
@@ -1,8 +1,8 @@
 public class HeapNode<Type> {
     public Type value;
-    public double priority;
+    public int priority;
 
-    public HeapNode(Type value, double priority) {
+    public HeapNode(Type value, int priority) {
         this.value = value;
         this.priority = priority;
     }
